+++ packages/router5/modules/plugins/browser/index.js
@@ -6,7 +6,9 @@ const defaultOptions = {
     forceDeactivate: true,
     useHash: false,
     hashPrefix: '',
-    base: false
+    base: false,
+    mergeState: false,
+    preserveHash: false
 };
 
 const source = 'popstate';
@@ -40,8 +42,15 @@ function browserPluginFactory(opts = {}, browser = safeBrowser) {
 
 
         function updateBrowserState(state, url, replace) {
-            if (replace) browser.replaceState(state, '', url);
-            else browser.pushState(state, '', url);
+            const trimmedState = state ? {
+                meta: state.meta,
+                name: state.name,
+                params: state.params,
+                path: state.path
+            } : state;
+            const finalState = options.mergeState === true ? { ...browser.getState(), ...trimmedState } : trimmedState;
+            if (replace) browser.replaceState(finalState, '', url);
+            else browser.pushState(finalState, '', url);
         }
 
         function onPopState(evt) {
@@ -102,7 +111,11 @@ function browserPluginFactory(opts = {}, browser = safeBrowser) {
             const historyState = browser.getState();
             const replace = opts.replace || fromState && router.areStatesEqual(toState, fromState, false) ||
                 opts.reload && historyState && router.areStatesEqual(toState, historyState, false);
-            updateBrowserState(toState, router.buildUrl(toState.name, toState.params), replace);
+            let url = router.buildUrl(toState.name, toState.params);
+            if(fromState === null && options.preserveHash === true) {
+                url += browser.getHash();
+            }
+            updateBrowserState(toState, url, replace);
         }
 
         return { onStart, onStop, onTransitionSuccess, onPopState };